Refactor server-client-related training platform code
